refactor : common 폴더 app 폴더 내부로 이동
