+++ src/app/layout.tsx
@@ -3,7 +3,7 @@ import { Provider } from '@/src/components/ui/provider';
 import { Theme } from '@chakra-ui/react';
 import localFont from 'next/font/local';
 import { Toaster } from '@/src/components/ui/toaster';
-import ErrorToast from '@/src/common/ErrorToast/ErrorToast';
+import ErrorToast from '@/src/app/common/ErrorToast/ErrorToast';
 export const metadata: Metadata = {
   title: 'JABBIT (재빗)',
   description: '재무관리의 모든 것',
